requirements.txt
apple-music-python==1.0.5
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
Django==3.2.9
django-model-utils==4.2.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
Pillow==8.4.0
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
